backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain[openai]>=1.1.0",
    "langchain-core>=1.1.0",
    "langgraph>=0.2.0",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "supabase>=2.24.0",
    "fastapi[standard]>=0.122.0",
    "pytest>=9.0.1",
    "requests>=2.31.0",
]
